+++ docs/changelog.rst
@@ -11,7 +11,8 @@ CHANGELOG
 
 **Minor changes**
 
-* MOBILE_TILES_URL settings is now a list which be used to merge different layers in zip mobil tiles
+* MOBILE_TILES_URL settings is now a list which be used to merge
+  different layers in tiles zip file for mobile application
 
 
 2.12.0 (2017-02-16)